refactor: move commands to menu
